backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.9
redis==5.1.1
httpx==0.27.2
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
websockets==12.0
